Enum better_term::Color

pub enum Color {

    Black,
    Red,
    Green,
    Yellow,
    Blue,
    Purple,
    Cyan,
    White,
    BrightBlack,
    BrightRed,
    BrightGreen,
    BrightYellow,
    BrightBlue,
    BrightPurple,
    BrightCyan,
    BrightWhite,
    Fixed(u8),
    RGB(u8, u8, u8),
    Hex(u32),

}

Default ansi colors

Example

use better_term::Color;

// prints Hello, world! in green and red
println!("{}Hello, {}world!", Color::BrightGreen, Color::BrightRed);

Variants

Black

Red

Green

Yellow

Blue

Purple

Cyan

White

BrightBlack

BrightRed

BrightGreen

BrightYellow

BrightBlue

BrightPurple

BrightCyan

BrightWhite

Fixed(u8)

Tuple Fields of Fixed

0: u8

RGB(u8, u8, u8)

Tuple Fields of RGB

0: u81: u82: u8

Hex(u32)

Tuple Fields of Hex

0: u32

Implementations

impl Color

pub fn to_style_fg(self) -> Style

Convert color to a style with the foreground set to this color

pub fn to_style_bg(self) -> Style

Convert color to a style with the background set to this color

pub fn hex_to_rgb(hex: u32) -> Color

convert a hexadecimal value to an RGB color value (hex example: 0x000000 for black)

pub fn as_fg(&self) -> String

Convert to the ansi value within a string (for output with the ansi char)

pub fn as_bg(&self) -> String

Convert to the ansi value within a string (for output with the ansi char